feat: style book
The cover of the book and the page were styled to resemble an ancient book.

The page got a parchment paper look.

The frontside of the book cover received an old looking book cover image and the logo with the name of the app.
The backside of the book cover and the backside of the book received a matching look.