feat: add navigation
On the backside of the cover a nagivation menu has been added to include the list of spells, a search function,
a filter function, to option to add one's own spell, to log in and to close the book.

An animation was added on hover for a more magical effect and prompt to click the
highlighted object.