feat: display preview spell list
When "Spells" is clicked a list of all public spells in the database is displayed
on the book page, showing their name, element and image.

Upon hover, an effect signals that the entries can be clicked.

The elements as well as the scrollbar have been styled to match the
branding.