feat: add register and login forms
When 'Login' is clicked, the login page renders on the side.

The user can input his email address and password to login.

If "Create Account" is clicked on the login page, the register page is rendered.

The user can input his first and last name, his email address and password to create an
account.

It is possible to return back to the login page by clicking "Back To Login" on the
register page (or the regular login option in the navigation menu).