feat: add search function
The user can now search for spells. When "Search" is clicked in the navigation menu,
the search page is rendered where the user can enter their search terms.
A list of spells is then rendered underneath if the search term is included in
their name. The user can then click on one of the search results to view the
spell in detail.