feat: add filter function
The user can now filter the spells by element. When 'Filter' is clicked in the
navigation menu, the filter interface is rendered on the page.
The user can then choose between the four elements and the corresponding
spells will be rendered underneath. The user may then click on one of the filter
results to view the spell in detail.